rename local agents package to gen_ai
